refactor: 모바일 키패드 관련 ux 개선
