+++ src/type.ts
@@ -1,3 +1,5 @@
+import { ChangeEvent, RefObject } from 'react';
+
 export type CardNumber = {
   first: string;
   second: string;
@@ -39,3 +41,33 @@ export type FormCardAddProps = {
 };
 
 export type actionName = 'CARDLIST_REQUEST' | 'CARDLIST_SUCCESS' | 'CARDLIST_FAILURE';
+
+type inputModeType =
+  | 'search'
+  | 'text'
+  | 'email'
+  | 'tel'
+  | 'url'
+  | 'none'
+  | 'numeric'
+  | 'decimal'
+  | undefined;
+
+export type InputProps = {
+  required: boolean;
+  inputType: string;
+  inputMode: inputModeType;
+  passwordType?: string;
+  className?: string;
+  value: string;
+  minDataLength: number;
+  maxDataLength: number;
+  name: string;
+  dataId: number;
+  refObject: RefObject<HTMLInputElement>;
+  handleError: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  nextFocus: (e: ChangeEvent) => void;
+  onFlip: () => void;
+  handleInputData: (index: number, e: ChangeEvent<HTMLInputElement>) => void;
+};
